+++ src/main/java/blackjack/domain/card/Result.java
@@ -2,5 +2,5 @@ package blackjack.domain.card;
 
 import java.util.Map;
 
-public record Result(Map<WinningResult, Integer> resultIntegerMap, double bettingResult) {
+public record Result(Map<WinningResult, Integer> resultIntegerMap, int bettingResult) {
 }
